rework Rejection system
The `Rejection` type can now nest and combine arbitrary `impl Reject`s,
so it is no longer bound to a small set of meanings. The ranking of
status codes is still used to determine which rejection gets priority.
A future enhancement should be to allow iterating over all the
rejections that were collected, so a certain `Filter::recover` could
change the priority if it desired.

- Adds `warp::reject::status(StatusCode)` to create a `Rejection` with
  any status code.
- Deprecates `Rejection::into_cause`, as it can no longer work. Always
  returns `Err(Rejection)`.
- Deprecates `warp::reject()`, since it current signals a `400 Bad
  Request`, but in newer versions, it will signal `404 Not Found`.
  It's deprecated simply to warn that the semantics are changing,
  but the function won't actually go away.